fix split packages with test-jars
