+++ modules/deployment-model-impl/src/test/java/org/mule/runtime/module/deployment/impl/internal/application/wrapper/TestApplicationWrapper.java
@@ -1,13 +1,14 @@
 /*
  * Copyright 2023 Salesforce, Inc. All rights reserved.
  */
-package org.mule.runtime.module.deployment.impl.internal.application;
+package org.mule.runtime.module.deployment.impl.internal.application.wrapper;
 
 import static org.mule.runtime.api.i18n.I18nMessageFactory.createStaticMessage;
 
 import org.mule.runtime.deployment.model.api.DeploymentException;
 import org.mule.runtime.deployment.model.api.DeploymentStopException;
 import org.mule.runtime.deployment.model.api.application.Application;
+import org.mule.runtime.module.deployment.impl.internal.application.ApplicationWrapper;
 
 import java.io.IOException;
