+++ modules/deployment-model-impl/src/test/java/org/mule/runtime/module/deployment/impl/internal/base/AbstractSplashScreenTestCase.java
@@ -1,7 +1,7 @@
 /*
  * Copyright 2023 Salesforce, Inc. All rights reserved.
  */
-package org.mule.runtime.module.deployment.impl.internal;
+package org.mule.runtime.module.deployment.impl.internal.base;
 
 import static org.junit.Assert.assertThat;
 import static org.mule.runtime.core.api.config.MuleProperties.MULE_HOME_DIRECTORY_PROPERTY;
